requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
beautifulsoup4==4.14.2
bs4==0.0.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.76.0
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
pycocotools==2.0.10
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.2
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
# Editable install with no version control (ssd-transfer==0.1.0)
-e c:\users\romai\documents\michelin\drive-download-20251021t194941z-1-002\ssd_transfer
stringzilla==4.2.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
